monitor_orders/requirements.txt
# This file lists the Python package dependencies for the sync_positions module.
requests
redis
pytz
mysql-connector-python

